Move all UI files into a top-level 'ui' folder.
