+++ src/index.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 
 import './index.css'
-import App from './components/app'
 import reportWebVitals from './reportWebVitals'
+import App from './ui/components/app'
 
 ReactDOM.render(
   <React.StrictMode>
