+++ src/ui/assets/fonts/NovaMono-Regular.ttf
