+++ src/ui/components/expedition-ended-screen.css
