+++ src/ui/components/log-panel.css
