+++ src/ui/components/map-panel.tsx
@@ -1,13 +1,14 @@
 import { ShaderSystem } from '@pixi/core'
 import { install } from '@pixi/unsafe-eval'
 import { Terrain } from 'db/terrain'
-import { FontNames } from 'fonts'
 import { isArray, noop } from 'lodash/fp'
 import * as PIXI from 'pixi.js'
 import { useCallback, useEffect, useRef, useState } from 'react'
 import { useRecoilState } from 'recoil'
-import { playerState } from 'state/player'
+import { playerState } from 'ui/state/player'
 import { World } from 'world/world'
+
+import { FontNames } from '../fonts'
 
 import { Panel } from './panel'
