+++ src/ui/components/panel.css
