+++ src/ui/components/panel.tsx
